Version 1.1 - Added Stores and Store-Tagging for gorcery items
Added the ability to add Stores and Tag grocery items with a store
so that you can see more easily where to get which grocery item.

Also reworked the UI to make it more usable on smartphones by
moving a lot of the functionality to the bottom of the screen
where they're easier to reach on larger phones.